feat: configure redux - restructure project
