+++ package.json
@@ -9,9 +9,11 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^2.3.0",
     "next": "14.2.13",
     "react": "^18",
     "react-dom": "^18",
+    "react-redux": "^9.1.2",
     "socket.io-client": "^4.8.0"
   },
   "devDependencies": {
