+++ src/hooks/store.ts
@@ -1,0 +1,5 @@
+import { useSelector, TypedUseSelectorHook, useDispatch } from 'react-redux';
+import type { AppDispatch, RootState } from '@/store';
+
+export const useAppDispatch = () => useDispatch<AppDispatch>();
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
